requirements.txt
django-bootstrap4==2.3.1
djangorestframework==3.12.1